docs(collapse): optimize collapse demo docs
