+++ examples/sites/demos/pc/app/collapse/before-close-composition-api.vue
@@ -1,5 +1,5 @@
 <template>
-  <tiny-collapse v-model="activeName" accordion :before-close="blockClose">
+  <tiny-collapse class="demo-collapse-wrap" v-model="activeName" accordion :before-close="beforeClose">
     <tiny-collapse-item title="一致性 Consistency" name="1">
       <div>与现实生活一致：与现实生活的流程、逻辑保持一致，遵循用户习惯的语言和概念；</div>
       <div>在界面中一致：所有的元素和结构需保持一致，比如：设计样式、图标和文本、元素的位置等。</div>
@@ -20,24 +20,22 @@
   </tiny-collapse>
 </template>
 
-<script>
-import { Collapse, CollapseItem, Modal } from '@opentiny/vue'
+<script setup>
+import { ref } from 'vue'
+import { Collapse as TinyCollapse, CollapseItem as TinyCollapseItem, Modal } from '@opentiny/vue'
 
-export default {
-  components: {
-    TinyCollapse: Collapse,
-    TinyCollapseItem: CollapseItem
-  },
-  data() {
-    return {
-      activeName: '1'
-    }
-  },
-  methods: {
-    blockClose(value) {
-      Modal.message({ message: '阻止面板关闭；传入value：' + value })
-      return false
-    }
-  }
+const activeName = ref('1')
+
+function beforeClose(item, activeNames) {
+  Modal.message({ message: '阻止面板切换，当前点击面板：' + item.name, status: 'error' })
+  return false
 }
 </script>
+
+<style scoped lang="less">
+.demo-collapse-wrap {
+  .tiny-collapse-item__content > * {
+    line-height: 1.8;
+  }
+}
+</style>
