--- examples/sites/demos/pc/app/collapse/custom-collapse-title.vue
@@ -1,105 +1,0 @@
-<template>
-  <tiny-collapse v-model="activeNames">
-    <tiny-collapse-item name="1">
-      <template #title>
-        <div class="title">
-          <div>一致性 Consistency</div>
-          <div @click.stop>
-            <a href="">查看监控数据</a>
-            <a href="">卸载</a>
-          </div>
-        </div>
-      </template>
-      <div>与现实生活一致：与现实生活的流程、逻辑保持一致，遵循用户习惯的语言和概念；</div>
-      <div>在界面中一致：所有的元素和结构需保持一致，比如：设计样式、图标和文本、元素的位置等。</div>
-    </tiny-collapse-item>
-    <tiny-collapse-item name="2" disabled>
-      <template #title>
-        <div class="title">
-          <div>效率 Efficiency</div>
-          <div @click.stop>
-            <a href="">查看监控数据</a>
-            <a href="">卸载</a>
-          </div>
-        </div>
-      </template>
-      <div>控制反馈：通过界面样式和交互动效让用户可以清晰的感知自己的操作；</div>
-      <div>页面反馈：操作后，通过页面元素的变化清晰地展现当前状态。</div>
-    </tiny-collapse-item>
-    <tiny-collapse-item title="效率 Efficiency" name="3">
-      <template #title-right>
-        <icon-clock-work class="tiny-icon-clockWork"></icon-clock-work>
-      </template>
-      <div>简化流程：设计简洁直观的操作流程；</div>
-      <div>清晰明确：语言表达清晰且表意明确，让用户快速理解进而作出决策；</div>
-      <div>帮助用户识别：界面简单直白，让用户快速识别而非回忆，减少用户记忆负担。</div>
-    </tiny-collapse-item>
-    <tiny-collapse-item title="可控 Controllability" name="4" disabled title-right="可控">
-      <div>用户决策：根据场景可给予用户操作建议或安全提示，但不能代替用户进行决策；</div>
-      <div>结果可控：用户可以自由的进行操作，包括撤销、回退和终止当前操作等。</div>
-    </tiny-collapse-item>
-  </tiny-collapse>
-</template>
-
-<script>
-import { Collapse, CollapseItem } from '@opentiny/vue'
-import { iconClockWork } from '@opentiny/vue-icon'
-
-export default {
-  components: {
-    TinyCollapse: Collapse,
-    TinyCollapseItem: CollapseItem,
-    IconClockWork: iconClockWork()
-  },
-  data() {
-    return {
-      activeNames: ['1']
-    }
-  }
-}
-</script>
-
-<style scoped>
-.icon-clockWork {
-  fill: var(--ti-common-color-line-active);
-  font-size: 18px;
-}
-
-.title {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.title a {
-  color: #1476ff;
-  position: relative;
-}
-
-.title a:hover {
-  font-weight: bold;
-}
-
-.title a:first-child::after {
-  content: '';
-  position: absolute;
-  height: 17px;
-  right: -12px;
-  border-right: 1px solid rgba(0, 0, 0, 0.15);
-}
-
-.title a:last-child {
-  margin-left: 24px;
-}
-
-.is-disabled a {
-  color: #c2c2c2;
-  pointer-events: none;
-}
-
-.tiny-icon-clockWork {
-  fill: var(--ti-common-color-line-active);
-  font-size: 18px;
-}
-</style>
