+++ examples/sites/demos/pc/app/collapse/disable.vue
@@ -1,15 +1,10 @@
 <template>
-  <tiny-collapse v-model="activeNames">
+  <tiny-collapse class="demo-collapse-wrap" v-model="activeNames">
     <tiny-collapse-item title="一致性 Consistency" name="1">
-      <template #icon>
-        <span>
-          <icon-delta-right class="icon-delta-right"></icon-delta-right>
-        </span>
-      </template>
       <div>与现实生活一致：与现实生活的流程、逻辑保持一致，遵循用户习惯的语言和概念；</div>
       <div>在界面中一致：所有的元素和结构需保持一致，比如：设计样式、图标和文本、元素的位置等。</div>
     </tiny-collapse-item>
-    <tiny-collapse-item title="反馈 Feedback" name="2">
+    <tiny-collapse-item title="反馈 Feedback" name="2" disabled>
       <div>控制反馈：通过界面样式和交互动效让用户可以清晰的感知自己的操作；</div>
       <div>页面反馈：操作后，通过页面元素的变化清晰地展现当前状态。</div>
     </tiny-collapse-item>
@@ -27,13 +22,11 @@
 
 <script>
 import { Collapse, CollapseItem } from '@opentiny/vue'
-import { IconDeltaRight } from '@opentiny/vue-icon'
 
 export default {
   components: {
     TinyCollapse: Collapse,
-    TinyCollapseItem: CollapseItem,
-    IconDeltaRight: IconDeltaRight()
+    TinyCollapseItem: CollapseItem
   },
   data() {
     return {
@@ -43,9 +36,10 @@ export default {
 }
 </script>
 
-<style scoped>
-.icon-delta-right {
-  fill: var(--ti-common-color-line-active);
-  font-size: 14px;
+<style scoped lang="less">
+.demo-collapse-wrap {
+  .tiny-collapse-item__content > * {
+    line-height: 1.8;
+  }
 }
 </style>
