+++ examples/sites/demos/pc/app/collapse/events.vue
@@ -1,11 +1,6 @@
 <template>
-  <tiny-collapse v-model="activeNames">
+  <tiny-collapse class="demo-collapse-wrap" v-model="activeNames" @change="change">
     <tiny-collapse-item title="一致性 Consistency" name="1">
-      <template #icon>
-        <span>
-          <tiny-icon-delta-right class="icon-delta-right"></tiny-icon-delta-right>
-        </span>
-      </template>
       <div>与现实生活一致：与现实生活的流程、逻辑保持一致，遵循用户习惯的语言和概念；</div>
       <div>在界面中一致：所有的元素和结构需保持一致，比如：设计样式、图标和文本、元素的位置等。</div>
     </tiny-collapse-item>
@@ -25,19 +20,34 @@
   </tiny-collapse>
 </template>
 
-<script setup>
-import { ref } from 'vue'
-import { Collapse as TinyCollapse, CollapseItem as TinyCollapseItem } from '@opentiny/vue'
-import { iconDeltaRight } from '@opentiny/vue-icon'
+<script>
+import { Collapse, CollapseItem, Modal } from '@opentiny/vue'
 
-const activeNames = ref(['1'])
-
-const TinyIconDeltaRight = iconDeltaRight()
+export default {
+  components: {
+    TinyCollapse: Collapse,
+    TinyCollapseItem: CollapseItem
+  },
+  data() {
+    return {
+      activeNames: ['1', '3']
+    }
+  },
+  methods: {
+    change(activeNames) {
+      Modal.message({
+        message: 'change 事件，当前激活面板：' + activeNames,
+        status: 'info'
+      })
+    }
+  }
+}
 </script>
 
-<style scoped>
-.icon-delta-right {
-  fill: var(--ti-common-color-line-active);
-  font-size: 14px;
+<style scoped lang="less">
+.demo-collapse-wrap {
+  .tiny-collapse-item__content > * {
+    line-height: 1.8;
+  }
 }
 </style>
