+++ examples/sites/demos/pc/app/collapse/icon-composition-api.vue
@@ -1,6 +1,9 @@
 <template>
   <tiny-collapse class="demo-collapse-wrap" v-model="activeNames">
     <tiny-collapse-item title="一致性 Consistency" name="1">
+      <template #icon>
+        <tiny-icon-delta-right class="icon-delta-right"></tiny-icon-delta-right>
+      </template>
       <div>与现实生活一致：与现实生活的流程、逻辑保持一致，遵循用户习惯的语言和概念；</div>
       <div>在界面中一致：所有的元素和结构需保持一致，比如：设计样式、图标和文本、元素的位置等。</div>
     </tiny-collapse-item>
@@ -20,20 +23,14 @@
   </tiny-collapse>
 </template>
 
-<script>
-import { Collapse, CollapseItem } from '@opentiny/vue'
+<script setup>
+import { ref } from 'vue'
+import { Collapse as TinyCollapse, CollapseItem as TinyCollapseItem } from '@opentiny/vue'
+import { iconDeltaRight } from '@opentiny/vue-icon'
 
-export default {
-  components: {
-    TinyCollapse: Collapse,
-    TinyCollapseItem: CollapseItem
-  },
-  data() {
-    return {
-      activeNames: ['1', '3']
-    }
-  }
-}
+const activeNames = ref(['1'])
+
+const TinyIconDeltaRight = iconDeltaRight()
 </script>
 
 <style scoped lang="less">
@@ -41,5 +38,10 @@ export default {
   .tiny-collapse-item__content > * {
     line-height: 1.8;
   }
+}
+
+.icon-delta-right {
+  fill: var(--ti-common-color-line-active);
+  font-size: 14px;
 }
 </style>
