+++ examples/sites/demos/pc/app/collapse/icon.vue
@@ -1,6 +1,11 @@
 <template>
-  <tiny-collapse v-model="activeNames">
+  <tiny-collapse class="demo-collapse-wrap" v-model="activeNames">
     <tiny-collapse-item title="一致性 Consistency" name="1">
+      <template #icon>
+        <span>
+          <icon-delta-right class="icon-delta-right"></icon-delta-right>
+        </span>
+      </template>
       <div>与现实生活一致：与现实生活的流程、逻辑保持一致，遵循用户习惯的语言和概念；</div>
       <div>在界面中一致：所有的元素和结构需保持一致，比如：设计样式、图标和文本、元素的位置等。</div>
     </tiny-collapse-item>
@@ -8,7 +13,7 @@
       <div>控制反馈：通过界面样式和交互动效让用户可以清晰的感知自己的操作；</div>
       <div>页面反馈：操作后，通过页面元素的变化清晰地展现当前状态。</div>
     </tiny-collapse-item>
-    <tiny-collapse-item title="效率 Efficiency" name="3" disabled>
+    <tiny-collapse-item title="效率 Efficiency" name="3">
       <div>简化流程：设计简洁直观的操作流程；</div>
       <div>清晰明确：语言表达清晰且表意明确，让用户快速理解进而作出决策；</div>
       <div>帮助用户识别：界面简单直白，让用户快速识别而非回忆，减少用户记忆负担。</div>
@@ -22,11 +27,13 @@
 
 <script>
 import { Collapse, CollapseItem } from '@opentiny/vue'
+import { IconDeltaRight } from '@opentiny/vue-icon'
 
 export default {
   components: {
     TinyCollapse: Collapse,
-    TinyCollapseItem: CollapseItem
+    TinyCollapseItem: CollapseItem,
+    IconDeltaRight: IconDeltaRight()
   },
   data() {
     return {
@@ -35,3 +42,16 @@ export default {
   }
 }
 </script>
+
+<style scoped lang="less">
+.demo-collapse-wrap {
+  .tiny-collapse-item__content > * {
+    line-height: 1.8;
+  }
+}
+
+.icon-delta-right {
+  fill: var(--ti-common-color-line-active);
+  font-size: 14px;
+}
+</style>
