+++ examples/sites/demos/pc/app/collapse/webdoc/collapse.en.md
@@ -1,7 +1,7 @@
 ---
-title: Collapse Panel
+title: Collapse
 ---
 
-# Collapse Panel
+# Collapse
 
 <div>You can specify dynamic pages or customized HTML files in the content area. The content area can be expanded and collapsed.</div>
